added light version for bot
